# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Twitter clone (Y)",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
      "UserRequest": {
        "type": "object",
        "properties": {
          "handle": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9]*$",
            "minLength": 3,
            "maxLength": 30
          }
        },
        "required": [
          "handle"
        ],
        "additionalProperties": false
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9]*$"
          },
          "handle": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9]*$",
            "minLength": 3,
            "maxLength": 30
          },
          "bookmarkedPosts": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 24,
              "maxLength": 24
            }
          },
          "following": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9]*$"
            }
          }
        },
        "required": [
          "_id",
          "handle",
          "bookmarkedPosts",
          "following"
        ],
        "additionalProperties": false
      },
      "PostRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 280
          },
          "replyTo": {
            "type": "string",
            "minLength": 24,
            "maxLength": 24
          }
        },
        "required": [
          "content"
        ],
        "additionalProperties": false
      },
      "PostResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 24,
            "maxLength": 24
          },
          "content": {
            "type": "string",
            "maxLength": 280
          },
          "replyTo": {
            "type": "string",
            "minLength": 24,
            "maxLength": 24
          },
          "authorId": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9]*$"
          },
          "likes": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9]*$"
            }
          },
          "comments": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 24,
              "maxLength": 24
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "content",
          "authorId",
          "likes",
          "comments",
          "createdAt",
          "lastModified"
        ],
        "additionalProperties": false
      },
      "CommentRequest": {
        "type": "object",
        "properties": {
          "post": {
            "type": "string",
            "minLength": 24,
            "maxLength": 24
          },
          "content": {
            "type": "string",
            "maxLength": 280
          }
        },
        "required": [
          "post",
          "content"
        ],
        "additionalProperties": false
      },
      "CommentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 24,
            "maxLength": 24
          },
          "content": {
            "type": "string",
            "maxLength": 280
          },
          "post": {
            "type": "string",
            "minLength": 24,
            "maxLength": 24
          },
          "authorId": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9]*$"
          },
          "likes": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9]*$"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "content",
          "post",
          "authorId",
          "likes",
          "createdAt",
          "lastModified"
        ],
        "additionalProperties": false
      }
    }
  },
  "paths": {
    "/comments": {
      "post": {
        "tags": [
          "Comments"
        ],
        "summary": "Create a new comment",
        "description": "Creates a new comment on a specific post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment created successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Comment created"
                }
              }
            }
          },
          "400": {
            "description": "Validation error (content too long, invalid post ID)"
          },
          "409": {
            "description": "Comment already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/comments/{id}": {
      "put": {
        "tags": [
          "Comments"
        ],
        "summary": "Update an existing comment by ID",
        "description": "Updates a comment's content and/or post reference using its unique ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The comment ID (MongoDB ObjectId)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comment updated successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Comment updated"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Comment not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "tags": [
          "Comments"
        ],
        "summary": "Get a comment by ID",
        "description": "Retrieves a single comment by its unique ID, including metadata like author and likes",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The comment ID (MongoDB ObjectId)"
          }
        ],
        "responses": {
          "200": {
            "description": "Comment found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Comment not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Comments"
        ],
        "summary": "Delete a comment by ID",
        "description": "Permanently deletes a comment from the database",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The comment ID (MongoDB ObjectId)"
          }
        ],
        "responses": {
          "200": {
            "description": "Comment deleted successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Comment deleted"
                }
              }
            }
          },
          "404": {
            "description": "Comment not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/posts": {
      "post": {
        "tags": [
          "Posts"
        ],
        "summary": "Create a new post",
        "description": "Creates a new post with content and optional reply reference",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Post created"
                }
              }
            }
          },
          "400": {
            "description": "Validation error (content too long, invalid replyTo ID)"
          },
          "409": {
            "description": "Post already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/posts/{id}": {
      "put": {
        "tags": [
          "Posts"
        ],
        "summary": "Update an existing post by ID",
        "description": "Updates a post's content and/or reply reference using its unique ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The post ID (MongoDB ObjectId)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post updated successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Post updated"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Post not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Get a post by ID",
        "description": "Retrieves a single post by its unique ID, including metadata like author, likes, and comments",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The post ID (MongoDB ObjectId)"
          }
        ],
        "responses": {
          "200": {
            "description": "Post found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "404": {
            "description": "Post not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Posts"
        ],
        "summary": "Delete a post by ID",
        "description": "Permanently deletes a post from the database",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The post ID (MongoDB ObjectId)"
          }
        ],
        "responses": {
          "200": {
            "description": "Post deleted successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Post deleted"
                }
              }
            }
          },
          "404": {
            "description": "Post not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a new user",
        "description": "Creates a new user with the given handle if it doesn't already exist",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "User created"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "User already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/{id}": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update an existing user by ID",
        "description": "Updates a user's information using their unique ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The user ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "User updated"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get a user by ID",
        "description": "Retrieves a single user by their unique ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The user ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete a user by ID",
        "description": "Permanently deletes a user from the database",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The user ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "User deleted"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "tags": []
}